refactor: rename cloud mongodb to mongodb
