+++ docs/data-sources/mongodb_image_product.md
