+++ examples/mongodb/variables.tf
