+++ internal/service/mongodb/mongodb_products_data_source.go
@@ -1,4 +1,4 @@
-package cloudmongodb
+package mongodb
 
 import (
 	"context"
