+++ internal/service/mongodb/mongodb_products_data_source_test.go
@@ -1,4 +1,4 @@
-package cloudmongodb_test
+package mongodb_test
 
 import (
 	"fmt"
